requirements.txt
# Core dependencies for Web Research Agent
google-generativeai>=0.8.0
requests>=2.32.0
beautifulsoup4>=4.12.0
html2text>=2024.2.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
